Move EventStore gRPC service to the package level
