+++ server/src/main/java/io/spine/server/event/GrpcService.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.event;
+
+import io.grpc.stub.StreamObserver;
+import io.spine.base.Event;
+import io.spine.base.Response;
+import io.spine.base.Responses;
+import io.spine.server.event.grpc.EventStoreGrpc;
+
+/**
+ * gRPC service over the locally running implementation of {@link EventStore}.
+ *
+ * @author Alexander Yevsyukov
+ */
+class GrpcService extends EventStoreGrpc.EventStoreImplBase {
+
+    private final LocalEventStore eventStore;
+
+    GrpcService(LocalEventStore eventStore) {
+        super();
+        this.eventStore = eventStore;
+    }
+
+    @Override
+    public void append(Event request, StreamObserver<Response> responseObserver) {
+        try {
+            eventStore.append(request);
+            responseObserver.onNext(Responses.ok());
+            responseObserver.onCompleted();
+        } catch (RuntimeException e) {
+            responseObserver.onError(e);
+        }
+    }
+
+    @Override
+    public void read(EventStreamQuery request, StreamObserver<Event> responseObserver) {
+        eventStore.read(request, responseObserver);
+    }
+}
